+++ app/src/js/components/Executions/executions-list.js
@@ -2,12 +2,12 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
-import { Link } from 'react-router-dom';
-import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
+import { Link, withRouter } from 'react-router-dom';
 import { listExecutionsByGranule } from '../../actions';
-import List from '../Table/Table';
 import { tableColumns } from '../../utils/table-config/executions-list';
+import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
+import List from '../Table/Table';
+import ExecutionSnapshot from './execution-snapshot';
 
 const breadcrumbConfig = [
   {
@@ -28,9 +28,11 @@ const ExecutionsList = ({
   match,
   executions
 }) => {
-  const { list } = executions || { list: { meta: { count: 0 }, results: [] } };
   const { params } = match || {};
   const { collectionId, granuleId } = params;
+  const { map } = executions || {};
+  const granuleExecutionslist = map[granuleId] || {};
+  const { meta } = granuleExecutionslist;
 
   const payload = {
     granules: [
@@ -40,6 +42,12 @@ const ExecutionsList = ({
       }
     ],
   };
+
+  function renderRowSubComponent(row) {
+    return (
+      <ExecutionSnapshot row={row} />
+    );
+  }
 
   return (
     <div className='page__component'>
@@ -62,13 +70,16 @@ const ExecutionsList = ({
         <section className="page__section">
           <div className="heading__wrapper--border">
             <h2 className="heading--medium heading--shared-content with-description">Total Executions
-              <span className='num-title'>{list.meta.count || 0}</span>
+              <span className='num-title'>{meta?.count || 0}</span>
             </h2>
           </div>
           <List
-            list={list}
+            list={granuleExecutionslist}
             tableColumns={tableColumns}
-            action={(() => listExecutionsByGranule(payload))}
+            action={() => listExecutionsByGranule(granuleId, payload)}
+            rowId='name'
+            initialSortId='updatedAt'
+            renderRowSubComponent={renderRowSubComponent}
           >
           </List>
         </section>