+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **CUMULUS-2462**
   - Reingest a granule from granule details page is updated to select a workflow execution arn for reingest.
   - Batch Reingest and Bulk Reingest Actions are updated to select the workflowName for reingest.
+- **CUMULUS-2585**
+  - Added failed execution step snapshot to execution list
 
 ### Changed